% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p4i_p4b_sector_technology_mapper}
\alias{p4i_p4b_sector_technology_mapper}
\title{P4i P4B technology and sector mapper}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28 rows and 4 columns.
}
\usage{
p4i_p4b_sector_technology_mapper
}
\description{
P4i P4B technology and sector mapper
}
\keyword{datasets}
