% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_score_function.R
\name{prep_scores}
\alias{prep_scores}
\title{Prepare data input for plotting aggregate climate scores}
\usage{
prep_scores(
  results_portfolio,
  overview = overview,
  scenario_source = "GECO2021"
)
}
\arguments{
\item{results_portfolio}{Data frame that contains pre-wrangled portfolio
level PACTA results from a PACTA for investors analysis.}

\item{overview}{Data frame that contains overview}

\item{scenario_source}{Character. Must be a
\code{scenario_source} featured in the \code{scenario_thresholds} data set.}
}
\value{
data.frame
}
\description{
Prepare data input for plotting aggregate climate scores based
on PACTA for investors output files. These files must have been wrangled with
\code{prep_data_executive_summary()} before they can be passed to this function.
}
